Move PD hidden state helpers to common
